moved old components and route checks
