+++ src/app/course/[courseId]/page.tsx
@@ -1,4 +1,4 @@
-import OfficeHourWrapper from "@components/OfficeHourWrapper";
+import OfficeHourWrapper from "@components/OldComponents/OfficeHourWrapper";
 import OfficeHourProvider from "@context/OfficeHourContext";
 import React from "react";
 interface PageProps {
